+++ tests/show/test_show_info_multiple.py
@@ -23,7 +23,7 @@ def get_connect_dict() -> Dict[str, Any]:
             return config_dict["database"]
 
 
-@pytest.mark.parametrize("show_id", [1083])
+@pytest.mark.parametrize("show_id", [1083, 1162])
 def test_show_info_retrieve_bluff_info_all(show_id: int):
     """Testing for :py:meth:`wwdtm.show.ShowInfoMultiple.retrieve_bluff_info_all`
 
@@ -180,15 +180,23 @@ def test_show_info_retrieve_guest_info_by_ids(show_ids: List[int]):
             ), f"'score' was not returned for the first guest for show ID {show_id}"
 
 
-@pytest.mark.parametrize("show_id", [1082])
-def test_show_info_retrieve_panelist_info_all(show_id: int):
+@pytest.mark.parametrize(
+    "show_id, include_decimal_scores", [(1082, True), (1082, False)]
+)
+def test_show_info_retrieve_panelist_info_all(
+    show_id: int, include_decimal_scores: bool
+):
     """Testing for :py:meth:`wwdtm.show.ShowInfoMultiple.retrieve_panelist_info_all`
 
     :param show_id: Show ID to test retrieving show panelist information
         for all shows retrieved
+    :param include_decimal_scores: Flag set to include decimal score columns
+        and values
     """
     info = ShowInfoMultiple(connect_dict=get_connect_dict())
-    shows_panelists = info.retrieve_panelist_info_all()
+    shows_panelists = info.retrieve_panelist_info_all(
+        include_decimal_scores=include_decimal_scores
+    )
 
     assert shows_panelists, "Panelist information for all shows could not be retrieved"
     assert (
@@ -202,72 +210,28 @@ def test_show_info_retrieve_panelist_info_all(show_id: int):
     assert (
         "score" in panelists[0]
     ), f"'score' was not returned for the first panelist for show ID {show_id}"
-
-
-@pytest.mark.parametrize("show_id", [1082])
-def test_show_info_retrieve_panelist_info_all_decimal(show_id: int):
-    """Testing for :py:meth:`wwdtm.show.ShowInfoMultiple.retrieve_panelist_info_all`
-    with decimal scores
-
-    :param show_id: Show ID to test retrieving show panelist information
-        for all shows retrieved
-    """
-    info = ShowInfoMultiple(connect_dict=get_connect_dict())
-    shows_panelists = info.retrieve_panelist_info_all(include_decimal_scores=True)
-
-    assert shows_panelists, "Panelist information for all shows could not be retrieved"
-    assert (
-        show_id in shows_panelists
-    ), f"Panelist information could not be retrieved for show ID {show_id}"
-
-    panelists = shows_panelists[show_id]
-    assert (
-        "id" in panelists[0]
-    ), f"'id' was not returned for the first panelist for show ID {show_id}"
-    assert (
-        "score" in panelists[0]
-    ), f"'score' was not returned for the first panelist for show ID {show_id}"
-
-
-@pytest.mark.parametrize("show_ids", [[1082, 1162]])
-def test_show_info_retrieve_panelist_info_by_ids(show_ids: List[int]):
-    """Testing for :py:meth:`wwdtm.show.ShowInfoMultiple.retrieve_panelist_info_by_ids`
-
-    :param show_ids: List of show IDs to test retrieving show panelist
-        information
-    """
-    info = ShowInfoMultiple(connect_dict=get_connect_dict())
-    shows_panelists = info.retrieve_panelist_info_by_ids(show_ids)
-
-    assert (
-        shows_panelists
-    ), f"Panelist information for show IDs {show_ids} could not be retrieved"
-
-    for show_id in shows_panelists:
+    if include_decimal_scores:
         assert (
-            show_id in shows_panelists
-        ), f"Panelist information could not be retrieved for show ID {show_id}"
-        panelists = shows_panelists[show_id]
-        if panelists:
-            assert (
-                "id" in panelists[0]
-            ), f"'id' was not returned for the first panelist for show ID {show_id}"
-            assert "score" in panelists[0], (
-                f"'score' was not returned for the first panelist for show ID "
-                f"{show_id}"
-            )
+            "score_decimal" in panelists[0]
+        ), f"'score_decimal' was not returned for the first panelist for show ID {show_id}"
 
 
-@pytest.mark.parametrize("show_ids", [[1082, 1162]])
-def test_show_info_retrieve_panelist_info_by_ids_decimal(show_ids: List[int]):
+@pytest.mark.parametrize(
+    "show_ids, include_decimal_scores", [([1082, 1162], True), ([1082, 1162], False)]
+)
+def test_show_info_retrieve_panelist_info_by_ids(
+    show_ids: List[int], include_decimal_scores: bool
+):
     """Testing for :py:meth:`wwdtm.show.ShowInfoMultiple.retrieve_panelist_info_by_ids`
 
     :param show_ids: List of show IDs to test retrieving show panelist
         information
+    :param include_decimal_scores: Flag set to include decimal score columns
+        and values
     """
     info = ShowInfoMultiple(connect_dict=get_connect_dict())
     shows_panelists = info.retrieve_panelist_info_by_ids(
-        show_ids, include_decimal_scores=True
+        show_ids, include_decimal_scores=include_decimal_scores
     )
 
     assert (
@@ -287,3 +251,7 @@ def test_show_info_retrieve_panelist_info_by_ids_decimal(show_ids: List[int]):
                 f"'score' was not returned for the first panelist for show ID "
                 f"{show_id}"
             )
+            if include_decimal_scores:
+                assert (
+                    "score_decimal" in panelists[0]
+                ), f"'score_decimal' was not returned for the first panelist for show ID {show_id}"