+++ wwdtm/show/info_multiple.py
@@ -42,27 +42,6 @@ class ShowInfoMultiple:
 
             self.database_connection = database_connection
 
-        try:
-            query = "SHOW COLUMNS FROM ww_showpnlmap WHERE Field = 'panelistlrndstart_decimal';"
-            cursor = self.database_connection.cursor()
-            cursor.execute(query)
-            start_decimal = cursor.fetchone()
-
-            query = (
-                "SHOW COLUMNS FROM ww_showpnlmap WHERE Field = 'panelistscore_decimal';"
-            )
-            cursor = self.database_connection.cursor()
-            cursor.execute(query)
-            score_decimal = cursor.fetchone()
-            cursor.close()
-
-            if start_decimal and score_decimal:
-                self.panelist_decimal_columns: bool = True
-            else:
-                self.panelist_decimal_columns: bool = False
-        except DatabaseError:
-            self.panelist_decimal_columns: bool = False
-
         self.utility = ShowUtility(database_connection=self.database_connection)
         self.loc_util = LocationUtility(database_connection=self.database_connection)
 
@@ -535,7 +514,7 @@ class ShowInfoMultiple:
             ranking information. If panelist information could not be
             retrieved, an empty list will be returned.
         """
-        if include_decimal_scores and self.panelist_decimal_columns:
+        if include_decimal_scores:
             query = """
                 SELECT s.showid AS show_id, pm.panelistid AS panelist_id,
                 p.panelist AS name, p.panelistslug AS slug,
@@ -619,7 +598,7 @@ class ShowInfoMultiple:
             if not valid_int_id(show_id):
                 return {}
 
-        if include_decimal_scores and self.panelist_decimal_columns:
+        if include_decimal_scores:
             query = """
                 SELECT s.showid AS show_id, pm.panelistid AS panelist_id,
                 p.panelist AS name, p.panelistslug AS slug,